flask-thsr/requirements.txt
# python-3.12

line-bot-sdk
flask
gunicorn
threading